添加练习2.6的代码

+++ src/lexer/Num.java
@@ -1,8 +1,8 @@
 package lexer;
 
 public class Num extends Token {
-	public final int value;
-	public Num(int v){
+	public final float value;
+	public Num(float v){
 		super(Tag.NUM);
 		value = v;
 	}
